% Generated by roxygen2 (4.0.1): do not edit by hand
\name{avg_event_to_travel}
\alias{avg_event_to_travel}
\title{A Cat Function}
\usage{
avg_event_to_travel(x)
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE.}
}
\description{
This function allows you to express your love of cats.
}
\examples{
avg_event_to_travel()
}
\keyword{cats}

